Added setColor to Agent and made getNeightbor take a radius as a parameter.

+++ src/simstation/Simulation.java
@@ -89,7 +89,7 @@ public class Simulation extends Model {
                 + "\nclock = " + clock);
     }
 
-    public synchronized Agent getNeighbor(Agent a, double radius) {
+    public synchronized Agent getNeighbor(Agent a, int radius) {
 
         double distance;
         double x;
